+++ src/cursor.cpp
@@ -74,7 +74,7 @@ namespace nogdb {
     auto cursor = metadata.begin() + currentIndex;
     auto recordDescriptor = *(cursor);
     auto classInfo = txn->_interface->schema()->getExistingClass(recordDescriptor.rid.first);
-    auto record = txn->_interface->record()->getRecord(classInfo, recordDescriptor);
+    auto record = txn->_interface->record()->getRecordWithBasicInfo(classInfo, recordDescriptor);
     record.setBasicInfo(DEPTH_PROPERTY, recordDescriptor._depth);
     result = Result{recordDescriptor, record};
     return true;
@@ -94,7 +94,7 @@ namespace nogdb {
     auto cursor = metadata.begin() + currentIndex;
     auto recordDescriptor = *(cursor);
     auto classInfo = txn->_interface->schema()->getExistingClass(recordDescriptor.rid.first);
-    auto record = txn->_interface->record()->getRecord(classInfo, recordDescriptor);
+    auto record = txn->_interface->record()->getRecordWithBasicInfo(classInfo, recordDescriptor);
     record.setBasicInfo(DEPTH_PROPERTY, recordDescriptor._depth);
     result = Result{recordDescriptor, record};
     return true;
@@ -121,7 +121,7 @@ namespace nogdb {
       auto cursor = metadata.begin();
       auto recordDescriptor = *(cursor);
       auto classInfo = txn->_interface->schema()->getExistingClass(recordDescriptor.rid.first);
-      auto record = txn->_interface->record()->getRecord(classInfo, recordDescriptor);
+      auto record = txn->_interface->record()->getRecordWithBasicInfo(classInfo, recordDescriptor);
       record.setBasicInfo(DEPTH_PROPERTY, recordDescriptor._depth);
       result = Result{recordDescriptor, record};
     }
@@ -136,7 +136,7 @@ namespace nogdb {
       auto cursor = metadata.end() - 1;
       auto recordDescriptor = *(cursor);
       auto classInfo = txn->_interface->schema()->getExistingClass(recordDescriptor.rid.first);
-      auto record = txn->_interface->record()->getRecord(classInfo, recordDescriptor);
+      auto record = txn->_interface->record()->getRecordWithBasicInfo(classInfo, recordDescriptor);
       record.setBasicInfo(DEPTH_PROPERTY, recordDescriptor._depth);
       result = Result{recordDescriptor, record};
     }
@@ -153,7 +153,7 @@ namespace nogdb {
     auto cursor = metadata.begin() + currentIndex;
     auto recordDescriptor = *(cursor);
     auto classInfo = txn->_interface->schema()->getExistingClass(recordDescriptor.rid.first);
-    auto record = txn->_interface->record()->getRecord(classInfo, recordDescriptor);
+    auto record = txn->_interface->record()->getRecordWithBasicInfo(classInfo, recordDescriptor);
     record.setBasicInfo(DEPTH_PROPERTY, recordDescriptor._depth);
     result = Result{recordDescriptor, record};
     return true;